+++ src/components/index.js
@@ -4,7 +4,8 @@ export { default as IBox } from './i-box.vue';
 export { default as IButton } from './i-button.vue';
 export { default as ICalendar } from './i-calendar.vue';
 export { default as ICheckbox } from './i-checkbox.vue';
-export { default as IDatepicker } from './i-datepicker.vue';
+export { default as IDatepicker } from './datepicker/i-datepicker.vue';
+export { default as IDateRangePicker } from './datepicker/i-date-range-picker.vue';
 export { default as IDialog } from './i-dialog.vue';
 export { default as IDropdown } from './dropdown/i-dropdown.vue';
 export { default as IDropdownOptions } from './dropdown/i-dropdown-options.vue';
